+++ class/plugin.js
@@ -182,6 +182,17 @@ class Plugin {
   }
 
   /**
+   * Force bind methods to self (```this```)
+   *
+   * @param {string[]} names - Method's names
+   */
+  selfBind (names) {
+    for (const name of names) {
+      this[name] = this[name].bind(this)
+    }
+  }
+
+  /**
    * Dispose internal references
    */
   dispose = () => {